+++ lib/src/widgets/custom_picker_location.dart
@@ -1,7 +1,4 @@
-import 'package:flutter/material.dart';
-
-import '../controller/picker_map_controller.dart';
-import '../osm_flutter.dart';
+part of osm_flutter;
 
 /// CustomPickerLocation
 ///